refactor(lsp): split lsp_integration.rs into shared TestClient + per-topic test files
Closes #2260
